Make same path copy/move a no-op

+++ src/TosAdapter.php
@@ -174,6 +174,10 @@ class TosAdapter implements FilesystemAdapter, PublicUrlGenerator, ChecksumProvi
 
     public function move(string $source, string $destination, Config $config): void
     {
+        if ($source === $destination) {
+            return;
+        }
+
         try {
             $this->copy($source, $destination, $config);
             $this->delete($source);
@@ -184,6 +188,10 @@ class TosAdapter implements FilesystemAdapter, PublicUrlGenerator, ChecksumProvi
 
     public function copy(string $source, string $destination, Config $config): void
     {
+        if ($source === $destination) {
+            return;
+        }
+
         try {
             /** @var string|null $visibility */
             $visibility = $config->get(Config::OPTION_VISIBILITY);
